+++ mentor/src/main/kotlin/com/sourceplusplus/mentor/job/RampDetectionMentor.kt
@@ -2,13 +2,24 @@ package com.sourceplusplus.mentor.job
 
 import com.sourceplusplus.mentor.MentorJob
 import com.sourceplusplus.mentor.MentorTask
-import com.sourceplusplus.mentor.task.*
+import com.sourceplusplus.mentor.task.analyze.CalculateLinearRegression
+import com.sourceplusplus.mentor.task.general.DelayTask
+import com.sourceplusplus.mentor.task.general.NoopTask
+import com.sourceplusplus.mentor.task.monitor.GetEndpoints
+import com.sourceplusplus.mentor.task.monitor.GetService
+import com.sourceplusplus.mentor.task.monitor.GetServiceInstance
+import com.sourceplusplus.mentor.task.monitor.GetTraces
 import com.sourceplusplus.protocol.artifact.trace.TraceOrderType
 import com.sourceplusplus.protocol.portal.QueryTimeFrame
 import io.vertx.core.Vertx
 
 /**
- * todo: description.
+ * Keeps track of endpoint durations for indications of 'The Ramp' [Smith and Williams 2002] performance anti-pattern.
+ * Uses a continuous linear regression model which requires a specific confidence before alerting developer.
+ * May also indicate root cause by searching trace stack for probable offenders.
+ *
+ * [Smith and Williams 2002] C. U. Smith and L. G. Williams, Performance Solutions: A Practical Guide to
+ * Creating Responsive, Scalable Software, Boston, MA, Addison-Wesley, 2002.
  *
  * @since 0.0.1
  * @author [Brandon Fergerson](mailto:bfergerson@apache.org)
@@ -26,6 +37,7 @@ class RampDetectionMentor(
                 GetService.SERVICE
             ),
 
+            //iterate endpoints (likely offenders more frequently than non-likely offenders)
             GetEndpoints(),
             GetTraces(
                 GetEndpoints.ENDPOINT_IDS,
@@ -33,17 +45,16 @@ class RampDetectionMentor(
                 timeFrame = QueryTimeFrame.LAST_15_MINUTES
             ),
 
-            //todo: ARIMA model?
-            CalculateLinearRegression(GetTraces.TRACE_RESULT, confidence),
+            //keep track of regression of endpoint duration
+            CalculateLinearRegression(GetTraces.TRACE_RESULT, confidence), //todo: ARIMA model?
+
+            //todo: search source code of endpoint for culprits
 
             if (config.repeatForever) {
                 DelayTask(config.repeatDelay)
             } else {
                 NoopTask()
             }
-
-            //todo: find endpoints with consistently increasing response time of a certain threshold
-            //todo: search source code of endpoint for culprits
         )
     }
 }
