+++ mentor/src/main/kotlin/com/sourceplusplus/mentor/task/monitor/GetTraceStacks.kt
@@ -1,4 +1,4 @@
-package com.sourceplusplus.mentor.task
+package com.sourceplusplus.mentor.task.monitor
 
 import com.sourceplusplus.mentor.MentorJob
 import com.sourceplusplus.mentor.MentorTask
