+++ mentor/src/test/kotlin/com/sourceplusplus/mentor/task/CalculateLinearRegressionTest.kt
@@ -3,6 +3,8 @@ package com.sourceplusplus.mentor.task
 import com.sourceplusplus.mentor.MentorJob
 import com.sourceplusplus.mentor.MentorTask
 import com.sourceplusplus.mentor.MentorTest
+import com.sourceplusplus.mentor.task.analyze.CalculateLinearRegression
+import com.sourceplusplus.mentor.task.monitor.GetTraces
 import com.sourceplusplus.protocol.artifact.trace.TraceOrderType
 import com.sourceplusplus.protocol.portal.QueryTimeFrame
 import io.vertx.core.Vertx
